+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 3.25.0-rc.2
 
+### Packaging
+
+* MapboxNavigationCore now requires [MapboxMaps v11.25.0-rc.2](https://github.com/mapbox/mapbox-maps-ios/releases/tag/v11.25.0-rc.2)
+* MapboxNavigationCore now requires [MapboxNavigationNative v324.25.0-rc.2](https://github.com/mapbox/mapbox-navigation-native-ios/releases/tag/v324.25.0-rc.2)
+
 ### Map
 
 * Fixed `NavigationMapView.automaticallySetDeclarativeMapContent` getter returning the inverse of the actual value.